ai-attention-backend/requirements.txt
flask
flask-cors
ultralytics
opencv-python-headless
requests
mysql-connector-python
matplotlib
